chore: change simulate to True
I don't want to write anything to the disk.

From the docs:

-s, --simulate                  Do not download the video and do not write
                                anything to disk
--no-simulate                   Download the video even if printing/listing
                                options are used
--skip-download                 Do not download the video but write all
                                related files (Alias: --no-download)

Also from yt_dlp/YoutubeDL.py:

simulate:  Do not download the video files. If unset (or None),
            simulate only if listsubtitles, listformats or list_thumbnails is used
skip_download:     Skip the actual download of the video file

So I think just using the 'simulate' is might be enough.